Keyboard shortcuts call urgent commands by default
Urgent commands mean that they are placed at the beginning of the
command queue, executed immediately. This is useful when a drawing
operation is taking a long time, and the user wishes to hide the turtle
to make it draw faster, or reset it immediately.

All drawing state commands have urgent counterparts now
(show!, hide!, clear!, pendown!, penup!, reset!).

Urgent commands formerly ended with a '_' in JavaScript, and were
aliased, ending with a '!' in Opal/Ruby. They are now renamed with a '!'
in JavaScript as well (accessed using square bracket notation), thus
Opal/Ruby aliases are no longer required.