Fix stop! command, fix push/pop
stop! did not actually reset the command queue.
In the last commit, addCommand! popped the last element off the
queue if a command was still executing, so perhaps this messes up
the array. This is now fixed by popping everything off.

push/pop now inserts a void vertex so that a line is not erroneously
drawn.